server/requirements.txt
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
h11==0.9.0
httptools==0.1.1
idna==2.9
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
starlette==0.13.2
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
